refactor: 모달 hook 분리 및 state를 RestaurantItem 에서 관리하도록 변경
